Move disable_qt_msghandler() to qtlog
